Rename `rustc_pattern_analysis/src/errors.rs` into `rustc_pattern_analysis/src/diagnostics.rs`
