+++ compiler/rustc_pattern_analysis/src/diagnostics.rs
